+++ DSCResources/MSFT_xExchMaintenanceMode/MSFT_xExchMaintenanceMode.psm1
@@ -297,8 +297,12 @@ function Set-TargetResource
                     $startDagScriptParams.Add("Force", 'true')
                 }
 
-                #Execute StartDagServerMaintenance.ps1
-                . $startDagServerMaintenanceScript @startDagScriptParams
+                # Execute StartDagServerMaintenance.ps1
+                Invoke-DotSourcedScript `
+                    -ScriptPath $startDagServerMaintenanceScript `
+                    -ScriptParams $startDagScriptParams `
+                    -SnapinsToRemove 'Microsoft.Exchange.Management.Powershell.E2010' `
+                    -Verbose:$VerbosePreference
             }
 
             #Set remaining components to offline
@@ -330,7 +334,16 @@ function Set-TargetResource
                 #doing the rest of the steps to take the server out of maintenance mode
                 try
                 {
-                    . $stopDagServerMaintenanceScript -serverName $env:COMPUTERNAME -Verbose
+                    $stopScriptParams = @{
+                        serverName = $env:COMPUTERNAME
+                        Verbose    = $true
+                    }
+
+                    Invoke-DotSourcedScript `
+                        -ScriptPath $stopDagServerMaintenanceScript `
+                        -ScriptParams $stopScriptParams `
+                        -SnapinsToRemove 'Microsoft.Exchange.Management.Powershell.E2010' `
+                        -Verbose:$VerbosePreference
                 }
                 catch
                 {
@@ -377,8 +390,6 @@ function Set-TargetResource
     {
         throw "Failed to retrieve maintenance mode status of server."
     }
-
-    Remove-HelperSnapin
 
     Remove-Item Alias:Write-Host -ErrorAction SilentlyContinue
 }
@@ -625,8 +636,6 @@ function Test-TargetResource
             }
         }
     }
-
-    Remove-HelperSnapin
 
     return $testResults
 }
@@ -1603,42 +1612,5 @@ function MoveActiveMailboxDatabase
     Move-ActiveMailboxDatabase @PSBoundParameters @moveDBParams
 }
 #endregion
-
-<#
-    .SYNOPSIS
-        Removes the Exchange PowerShell snapin, which is loaded by the
-        Start/StopDagMaintennace.ps1 scripts in the $Exscripts
-        directory. Prevents an issue where if a snapin is added by multiple
-        modules during the same session, subsequent additions of the same
-        module fail with 'An item with the same key has already been added'.
-
-    .NOTES
-        This similar function exists in the files
-        MSFT_xExchAntiMalwareScanning.psm1 and MSFT_xExchMaintenanceMode.psm1.
-        This was initially attempted to be put in xExchangeHelper.psm1 instead.
-        However when xExchangeHelper.psm1 is loaded as a NestedModule from
-        xExchange.psd1, functions within xExchangeHelper.psm1 do not appear to
-        be able to detect added snapins loaded by scripts called from other
-        modules. The added snapins were only detectable when running
-        Get-PSSnapin directly from functions within
-        MSFT_xExchAntiMalwareScanning.psm1 and MSFT_xExchMaintenanceMode.psm1.
-#>
-function Remove-HelperSnapin
-{
-    [CmdletBinding()]
-    param()
-
-    $snapinsToRemove = @('Microsoft.Exchange.Management.Powershell.E2010')
-
-    foreach ($snapin in $snapinsToRemove)
-    {
-        if ($null -ne (Get-PSSnapin -Name $snapin -ErrorAction SilentlyContinue))
-        {
-            Write-Verbose -Message "'$snapin' snapin is currently loaded. Removing."
-
-            Remove-PSSnapin -Name $snapin -ErrorAction SilentlyContinue -Confirm:$false
-        }
-    }
-}
 
 Export-ModuleMember -Function *-TargetResource