+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update appveyor.yml to use the default template.
 - Added default template files .codecov.yml, .gitattributes, and .gitignore, and
   .vscode folder.
+- Add Unit Tests for xExchAntiMalwareScanning
 
 ## 1.23.0.0